Remove unnecessary processing with DB and add new filtering logic.
