+++ profit-system/src/components/pages/CompaniesList.tsx
@@ -1,5 +1,5 @@
 import { Box } from "@chakra-ui/react";
-import { memo, FC, useState, ChangeEvent } from "react";
+import { memo, FC, useState, ChangeEvent, useEffect } from "react";
 import { HeadLine } from "../atoms/HeadLine";
 import { MainScreenTopContainer } from "../molecules/container/MainScreenTopContainer";
 import { ContentBgTemplate } from "../molecules/container/ContentBgTemplateContainer";
@@ -8,12 +8,13 @@ import { CompaniesTableTemplateList } from "../molecules/list/CompaniesTableTemp
 import { IconButton } from "../atoms/button/IconButton";
 import { useNavigate } from "react-router-dom";
 import { db } from "../../firebase";
-import { collection, getDocs, query, where } from "firebase/firestore";
-import { CompanyDbType } from "../../types/company/company";
+import { collection, getDocs } from "firebase/firestore";
+import { CompanyDbType } from "../../types/company/CompanyDbType";
 
 export const CompaniesList: FC = memo(() => {
   const navigate = useNavigate();
   const [searchValue, setSearchValue] = useState("");
+  const [companyData, setCompanyData] = useState<CompanyDbType[]>([]);
   const [filteredCompanyList, setFilteredCompanyList] = useState<
     CompanyDbType[]
   >([]);
@@ -22,24 +23,32 @@ export const CompaniesList: FC = memo(() => {
     setSearchValue(e.target.value);
   };
 
-  const onClickFilter = async () => {
-    const filteredCompanyListData: CompanyDbType[] = [];
-    const q = query(
-      collection(db, "company"),
-      where("name", ">=", searchValue),
-      where("name", "<", searchValue + "\uf8ff")
-    );
-    const querySnapshot = await getDocs(q);
-    querySnapshot.forEach((doc) => {
-      filteredCompanyListData.push({
-        id: doc.id,
-        name: doc.data().name,
-        address: doc.data().address,
-        postCode: doc.data().postCode,
-        phone: doc.data().phone,
+  useEffect(() => {
+    const getCompanyData = async () => {
+      const companyList: CompanyDbType[] = [];
+      const querySnapshot = await getDocs(collection(db, "company"));
+      querySnapshot.forEach((doc) => {
+        companyList.push({
+          id: doc.id,
+          name: doc.data().name,
+          address: doc.data().address,
+          postCode: doc.data().postCode,
+          phone: doc.data().phone,
+        });
       });
+      setCompanyData(companyList);
+      setFilteredCompanyList(companyList);
+    };
+    getCompanyData();
+  }, []);
+
+  const onClickFilter = () => {
+    const filteredCompany = companyData.filter((company) => {
+      if (company.name.indexOf(searchValue) !== -1) {
+        return company;
+      }
     });
-    setFilteredCompanyList(filteredCompanyListData);
+    setFilteredCompanyList(filteredCompany);
   };
 
   const onClickRegisterPage = () => {
